requirements.txt
ccxt==4.4.8
pandas==2.2.3
numpy==2.2.5
aiogram==3.19.0
python-dotenv==1.0.0
openpyxl==3.1.2
pytest==7.4.3
asyncio==3.4.3
winloop==0.5.1